refactor(spa-editor): replace contexts barrel imports with direct module paths
CodeRabbit flagged the four remaining barrel reads after the deslap.
Direct imports (`./contexts/garmin-bridge-context`, `./contexts/ThemeContext`,
`./contexts/analytics-context`) keep tree-shaking and refactors easier
to reason about.
